Move data parsers into a subpackage
This patch groups our two parsers into a subpackage, as well as group them into a subcommand in main.py
