+++ .travis.yml
@@ -40,7 +40,7 @@ before_script:
 script:
   - pylint -j $(nproc) anyway tests
   - eslint static/js --ignore-path=static/js/.eslintignore
-  - ./main.py process_data
-  - ./main.py import_united_data --light
+  - ./main.py process cbs
+  - ./main.py process united --light
   - pytest tests -m "not browser"
   - if [ "$TRAVIS_PULL_REQUEST" == "false" ]; then for BROWSER in Chrome Firefox; do pytest tests -m browser --driver SauceLabs --capability browserName $BROWSER --capability tunnelIdentifier $TRAVIS_JOB_NUMBER; done; fi
