+++ anyway/parsers/cbs.py
@@ -10,10 +10,10 @@ from datetime import datetime
 from flask.ext.sqlalchemy import SQLAlchemy
 from sqlalchemy import or_, and_
 
-from . import field_names, localization
-from .models import Marker, Involved, Vehicle
-from . import models
-from .utilities import ItmToWGS84, init_flask, CsvReader, time_delta
+from .. import field_names, localization
+from ..models import Marker, Involved, Vehicle
+from .. import models
+from ..utilities import ItmToWGS84, init_flask, CsvReader, time_delta
 import logging
 
 # Headless servers cannot use GUI file dialog and require raw user input
